commit 2

+++ compute_weight.cpp
@@ -4,8 +4,8 @@
 
 #include "compute_weight.h"
 
-vector<float> compute_weight(const Point2f &p, const Point2f &K1, const Point2f &K2) {
-    vector<float> r(2);
+vector<double> compute_weight(const Point2d &p, const Point2d &K1, const Point2d &K2) {
+    vector<double> r(2);
     if (p.x < K1.x) {
         r[0] = 0;
         r[1] = 1;
@@ -18,9 +18,9 @@ vector<float> compute_weight(const Point2f &p, const Point2f &K1, const Point2f 
         return r;
     }
 
-    float a = (p.x - K1.x) * (K2.x - K1.x);
-    float b = (p.y - K1.y) * (K2.y - K1.y);
-    float c = powf(K2.x - K1.x,2) + powf(K2.y - K1.y,2);
+    double a = (p.x - K1.x) * (K2.x - K1.x);
+    double b = (p.y - K1.y) * (K2.y - K1.y);
+    double c = pow(K2.x - K1.x,2) + pow(K2.y - K1.y,2);
 
     r[0] = abs(a + b)/c;
     r[1] = 1 - r[0];
