+++ compute_weight.h
@@ -11,6 +11,6 @@
 using namespace std;
 using namespace cv;
 
-vector<float> compute_weight(const Point2f &p, const Point2f &K1, const Point2f &K2);
+vector<double> compute_weight(const Point2d &p, const Point2d &K1, const Point2d &K2);
 
 #endif //AANAP_COMPUTE_WEIGHT_H
