+++ homography_warp.h
@@ -10,8 +10,8 @@
 using namespace cv;
 using namespace std;
 
-Point2f homography_warp(const Mat& src, const Mat& H, const Point2f offset, const Size s, Mat& dst);
+Point2d homography_warp(const Mat& src, const Mat& H, const Point2d offset, const Size s, Mat& dst);
 
-Mat homography_linearization(const Mat &H, const Point &center_point, const vector<Point2f> &anchor_points);
+Mat homography_linearization(const Mat &H, const Point &center_point, const vector<Point2d> &anchor_points);
 
 #endif //AANAP_HOMOGRAPHY_WARP_H
